requirements.txt
PyGithub
getpass